feat(cantidad): unidades enteras en todo el sistema (ventas, compras, traslados, inventario)
Pedido del cliente: los productos se venden por unidades enteras, nunca
fraccionadas. Antes cantidad era NUMERIC(12,3) en las 4 tablas (y
existencia_actual), con un comentario explicito en market-flutter
justificando decimales para productos por peso/volumen que nunca aplico a
este negocio.

Backend: columnas de BD a NUMERIC(12,0) via migraciones Liquibase con
precondicion HALT (nunca redondear en silencio una cantidad fraccionaria
real si algun dia existiera); @Digits(fraction=0) en los 4 DTOs de request;
chequeo de dominio (cantidad.remainder(ONE) != 0) en LineaVenta/LineaCompra/
LineaTraslado/MovimientoInventario, mismo estilo que el chequeo de "> 0" ya
existente.

Frontend (backoffice): los 4 inputs de cantidad pasan de step="0.001" a
step="1" min="1" (validacion nativa del form).

Flutter: LineaCarrito.cantidad de Decimal a int -- sin migracion de Isar,
ese campo ya se persistia como String. El carrito solo permitia +/-1 via
stepper, nunca entrada decimal libre, asi que el comportamiento real no
cambia para el usuario.

mvn test: 94 clases sin fallos. pnpm typecheck/test: limpio. flutter
analyze/test: limpio (62 tests).

+++ market-flutter/lib/core/db/local_store_io.dart
@@ -430,7 +430,7 @@ class IsarLocalStore implements LocalStore {
               productoId: l.productoId,
               nombre: l.nombre,
               precioUnitario: Decimal.parse(l.precioUnitario),
-              cantidad: Decimal.parse(l.cantidad),
+              cantidad: int.parse(l.cantidad),
             ),
           )
           .toList(),
